refactor: changed to kebab case for consistency
