+++ packages/api-key-service/.gitignore
@@ -25,4 +25,4 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 
-publicKeys.json
+public-keys.json
