+++ packages/api-key-service/src/functions/paginate.ts
@@ -2,7 +2,7 @@ import { ApiKeyModel } from '@/models'
 import { httpHandler, UnauthorizedException } from '@homeservenow/serverless-aws-handler'
 import { batchGetApiKeys } from './../services'
 import { connectSessionVerifyDecodeIdTokenWithPublicKeys, LoginIdentity } from '@reapit/connect-session'
-import publicKeys from './../../publicKeys.json'
+import publicKeys from './../../public-keys.json'
 
 type Pagintation<T> = {
   items: T[]
