+++ packages/api-key-service/src/functions/update-api-key.ts
@@ -10,7 +10,7 @@ import { ApiKeyDto } from '@/dto'
 import { ApiKeyModel } from '@/models'
 import { getApiKey, updateApiKey as update } from '@/services'
 import { connectSessionVerifyDecodeIdTokenWithPublicKeys, LoginIdentity } from '@reapit/connect-session'
-import publicKeys from '../../publicKeys.json'
+import publicKeys from '../../public-keys.json'
 
 export const updateApiKey = httpHandler<ApiKeyDto, ApiKeyModel>({
   handler: async ({ body, event }): Promise<ApiKeyModel> => {
