+++ packages/utils/src/node/fetch-public-keys/constants.js
@@ -1,4 +1,4 @@
-const publicKeyFileName = 'publicKeys.json'
+const publicKeyFileName = 'public-keys.json'
 const configFileName = 'config.json'
 
 module.exports = {
